Split Card component into Header, Body, Footer
