+++ app/javascript/src/components/Dashboard/Notes/Card/Body.jsx
@@ -1,0 +1,15 @@
+import React from "react";
+
+import { Typography } from "neetoui/v2";
+
+const Body = ({ note }) => {
+  return (
+    <>
+      <Typography style="body2" className="text-gray-500">
+        {note.description}
+      </Typography>
+      <hr className="my-3" />
+    </>
+  );
+};
+export default Body;
